pyproject.toml
[tool.poetry]
name = "manage-aiworkforce"
version = "0.1.0"
description = "Manage your Relevance AI workforce and deployments via filestores and GitHub."
authors = ["Alex Jenkins <alex@relevanceai.com>"]
readme = "README.md"
packages = [
    { include = "aiworkforce" },
    { include = "agentsmith" }
]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pytz = "^2025.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
